Updated partition_setup according to latest version

+++ partition_setup_dev.sh
@@ -13,17 +13,31 @@ mount "$part1" /tmp/1
 mount "$part2" /tmp/2
 mount "$part3" /tmp/3
 
-sed /tmp/1/cmdline.txt -i -e "s|root=/dev/[^ ]*|root=${part2}|"
-sed /tmp/2/etc/fstab -i -e "s|^.* / |${part2}  / |"
-sed /tmp/2/etc/fstab -i -e "s|^.* /boot |${part1}  /boot |"
+sed /tmp/1/cmdline.txt -i -e "s|root=[^ ]*|root=${part2}|"
+
+sed /tmp/2/etc/fstab -i -e "s|^.* / |${part2}  / "
+
+sed /tmp/2/etc/fstab -i -e "s|^.* /boot |${part1}  /boot "
+
+if [ -f /mnt/ssh ]; then
+  cp /mnt/ssh /tmp/1/
+fi
+
+if [ -f /mnt/ssh.txt ]; then
+  cp /mnt/ssh.txt /tmp/1/
+fi
+
+if [ -f /settings/wpa_supplicant.conf ]; then
+  cp /settings/wpa_supplicant.conf /tmp/1/
+  cp /settings/wpa_supplicant.conf /tmp/2/etc/wpa_supplicant
+fi
 
 if ! grep -q resize /proc/cmdline; then
-  sed -i 's/ quiet init=.*$//' /tmp/1/cmdline.txt
+  sed -i 's| quiet init=/usr/lib/raspi-config/init_resize.sh||' /tmp/1/cmdline.txt
 fi
 
 rm /tmp/2/etc/init.d/apply_noobs_os_config
 cp /tmp/3/apply_noobs_os_config /tmp/2/etc/init.d
-cp /settings/wpa_supplicant.conf /tmp/2/etc/wpa_supplicant
 
 rm -r /settings/cache/*
 umount /tmp/1
